Remove ineffective tests
